# api/docs/swagger.yaml
basePath: /
definitions:
  httpapi.authResponse:
    properties:
      token:
        type: string
      user:
        $ref: '#/definitions/httpapi.userResponse'
    type: object
  httpapi.chatResponse:
    properties:
      created_at:
        type: string
      id:
        type: string
      invite_code:
        type: string
      name:
        type: string
      owner_id:
        type: string
    type: object
  httpapi.createChatRequest:
    properties:
      name:
        type: string
    type: object
  httpapi.errorResponse:
    properties:
      error:
        type: string
    type: object
  httpapi.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  httpapi.messageResponse:
    properties:
      chat_id:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      user_id:
        type: string
    type: object
  httpapi.registerRequest:
    properties:
      display_name:
        type: string
      email:
        type: string
      password:
        type: string
    type: object
  httpapi.userResponse:
    properties:
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: WebSocket messenger API
  title: ws-messenger API
  version: 0.1.0
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные логина
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/httpapi.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Логин
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные регистрации
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/httpapi.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      summary: Регистрация
      tags:
      - auth
  /chats:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpapi.chatResponse'
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Список чатов пользователя
      tags:
      - chats
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные чата
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/httpapi.createChatRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httpapi.chatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Создать чат
      tags:
      - chats
  /chats/{id}/messages:
    get:
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: RFC3339 timestamp
        in: query
        name: before
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/httpapi.messageResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: История сообщений
      tags:
      - messages
  /chats/join/{code}:
    post:
      parameters:
      - description: Invite code
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httpapi.chatResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httpapi.errorResponse'
      security:
      - BearerAuth: []
      summary: Вступить в чат по инвайт-коду
      tags:
      - chats
  /ws/chats/{id}:
    get:
      parameters:
      - description: Chat ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "101":
          description: Switching Protocols
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: WebSocket подключение
      tags:
      - ws
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"